backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-huggingface
openai
pinecone
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
aiohttp
requests